+++ CHANGELOG.md
@@ -28,7 +28,7 @@ and this project follows something close to [Semantic Versioning](https://semver
 
 ### Added
 - For nsx files, within a single notebook now compare an attachment md5 and file name to those already written to disk and if the file is a duplicate it is not saved to disk and links will be to the existing file.
-
+- Default export folder `notes` and default attachement folder `attachements` added to quick settings, previously it would use what was in the config.ini.
 ### Fixed
 - Checklist items in HTML/NSX on the same line will now be converted
 - Markdown to HTML conversions - checklist items are now enabled in the exported html.  Previously the status was set to disabled and were not clickable.